simplify since DVHA-MLCA created with cmd-line UI, fx-weighted complexity scores Previously, plans with multiple fraction groups reported complexity scores as straight sums. They are now weighted by fxs over each Rx for the plan.
